// src/database/types.ts
export type AuthAalLevel = "aal1" | "aal2" | "aal3";

export type AuthCodeChallengeMethod = "plain" | "s256";

export type AuthFactorStatus = "unverified" | "verified";

export type AuthFactorType = "phone" | "totp" | "webauthn";

export type AuthOneTimeTokenType = "confirmation_token" | "email_change_token_current" | "email_change_token_new" | "phone_change_token" | "reauthentication_token" | "recovery_token";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Int8 = ColumnType<string, bigint | number | string, bigint | number | string>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type OwnerType = "contributor" | "creator" | "viewer";

export type PeriodType = "annually" | "daily" | "monthly" | "weekly";

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export type TransactionAction = "create" | "delete" | "update";

export type UserType = "admin" | "user";

export type WalletType = "budget" | "savings";

export interface AuthAuditLogEntries {
  created_at: Timestamp | null;
  id: string;
  instance_id: string | null;
  ip_address: Generated<string>;
  payload: Json | null;
}

export interface AuthFlowState {
  auth_code: string;
  auth_code_issued_at: Timestamp | null;
  authentication_method: string;
  code_challenge: string;
  code_challenge_method: AuthCodeChallengeMethod;
  created_at: Timestamp | null;
  id: string;
  provider_access_token: string | null;
  provider_refresh_token: string | null;
  provider_type: string;
  updated_at: Timestamp | null;
  user_id: string | null;
}

export interface AuthIdentities {
  created_at: Timestamp | null;
  /**
   * Auth: Email is a generated column that references the optional email property in the identity_data
   */
  email: Generated<string | null>;
  id: Generated<string>;
  identity_data: Json;
  last_sign_in_at: Timestamp | null;
  provider: string;
  provider_id: string;
  updated_at: Timestamp | null;
  user_id: string;
}

export interface AuthInstances {
  created_at: Timestamp | null;
  id: string;
  raw_base_config: string | null;
  updated_at: Timestamp | null;
  uuid: string | null;
}

export interface AuthMfaAmrClaims {
  authentication_method: string;
  created_at: Timestamp;
  id: string;
  session_id: string;
  updated_at: Timestamp;
}

export interface AuthMfaChallenges {
  created_at: Timestamp;
  factor_id: string;
  id: string;
  ip_address: string;
  otp_code: string | null;
  verified_at: Timestamp | null;
  web_authn_session_data: Json | null;
}

export interface AuthMfaFactors {
  created_at: Timestamp;
  factor_type: AuthFactorType;
  friendly_name: string | null;
  id: string;
  last_challenged_at: Timestamp | null;
  phone: string | null;
  secret: string | null;
  status: AuthFactorStatus;
  updated_at: Timestamp;
  user_id: string;
  web_authn_aaguid: string | null;
  web_authn_credential: Json | null;
}

export interface AuthOneTimeTokens {
  created_at: Generated<Timestamp>;
  id: string;
  relates_to: string;
  token_hash: string;
  token_type: AuthOneTimeTokenType;
  updated_at: Generated<Timestamp>;
  user_id: string;
}

export interface AuthRefreshTokens {
  created_at: Timestamp | null;
  id: Generated<Int8>;
  instance_id: string | null;
  parent: string | null;
  revoked: boolean | null;
  session_id: string | null;
  token: string | null;
  updated_at: Timestamp | null;
  user_id: string | null;
}

export interface AuthSamlProviders {
  attribute_mapping: Json | null;
  created_at: Timestamp | null;
  entity_id: string;
  id: string;
  metadata_url: string | null;
  metadata_xml: string;
  name_id_format: string | null;
  sso_provider_id: string;
  updated_at: Timestamp | null;
}

export interface AuthSamlRelayStates {
  created_at: Timestamp | null;
  flow_state_id: string | null;
  for_email: string | null;
  id: string;
  redirect_to: string | null;
  request_id: string;
  sso_provider_id: string;
  updated_at: Timestamp | null;
}

export interface AuthSchemaMigrations {
  version: string;
}

export interface AuthSessions {
  aal: AuthAalLevel | null;
  created_at: Timestamp | null;
  factor_id: string | null;
  id: string;
  ip: string | null;
  /**
   * Auth: Not after is a nullable column that contains a timestamp after which the session should be regarded as expired.
   */
  not_after: Timestamp | null;
  refreshed_at: Timestamp | null;
  tag: string | null;
  updated_at: Timestamp | null;
  user_agent: string | null;
  user_id: string;
}

export interface AuthSsoDomains {
  created_at: Timestamp | null;
  domain: string;
  id: string;
  sso_provider_id: string;
  updated_at: Timestamp | null;
}

export interface AuthSsoProviders {
  created_at: Timestamp | null;
  id: string;
  /**
   * Auth: Uniquely identifies a SSO provider according to a user-chosen resource ID (case insensitive), useful in infrastructure as code.
   */
  resource_id: string | null;
  updated_at: Timestamp | null;
}

export interface AuthUsers {
  aud: string | null;
  banned_until: Timestamp | null;
  confirmation_sent_at: Timestamp | null;
  confirmation_token: string | null;
  confirmed_at: Generated<Timestamp | null>;
  created_at: Timestamp | null;
  deleted_at: Timestamp | null;
  email: string | null;
  email_change: string | null;
  email_change_confirm_status: Generated<number | null>;
  email_change_sent_at: Timestamp | null;
  email_change_token_current: Generated<string | null>;
  email_change_token_new: string | null;
  email_confirmed_at: Timestamp | null;
  encrypted_password: string | null;
  id: string;
  instance_id: string | null;
  invited_at: Timestamp | null;
  is_anonymous: Generated<boolean>;
  /**
   * Auth: Set this column to true when the account comes from SSO. These accounts can have duplicate emails.
   */
  is_sso_user: Generated<boolean>;
  is_super_admin: boolean | null;
  last_sign_in_at: Timestamp | null;
  phone: Generated<string | null>;
  phone_change: Generated<string | null>;
  phone_change_sent_at: Timestamp | null;
  phone_change_token: Generated<string | null>;
  phone_confirmed_at: Timestamp | null;
  raw_app_meta_data: Json | null;
  raw_user_meta_data: Json | null;
  reauthentication_sent_at: Timestamp | null;
  reauthentication_token: Generated<string | null>;
  recovery_sent_at: Timestamp | null;
  recovery_token: string | null;
  role: string | null;
  updated_at: Timestamp | null;
}

export interface ConstantsCategories {
  id: Generated<number>;
  name: string;
}

export interface ConstantsCurrencies {
  code: string;
  id: Generated<number>;
  name: string;
}

export interface ProsperBudgetMetadata {
  assigned_amount: Int8;
  category_id: number;
  created_at: Generated<Timestamp>;
  id: Int8;
  name: string;
  period: PeriodType;
  preferred_currency_id: number;
  refreshed_at: Generated<Timestamp>;
  updated_at: Generated<Timestamp>;
}

export interface ProsperPartnerInvitations {
  expiration: Timestamp;
  id: Int8;
  nonce: Int8;
  receiver_id: Int8;
  sender_id: Int8;
  type: WalletType;
  wallet_id: Int8;
}

export interface ProsperProfiles {
  auth_id: string;
  created_at: Generated<Timestamp>;
  id: string;
  name: string;
  profile_url: string | null;
  type: Generated<UserType>;
  updated_at: Generated<Timestamp>;
  username: string;
}

export interface ProsperSavingsMetadata {
  category_id: number;
  created_at: Generated<Timestamp>;
  id: Int8;
  name: string;
  preferred_currency_id: number;
  priority: Generated<boolean | null>;
  target_amount: Int8;
  target_date: Timestamp;
  updated_at: Generated<Timestamp>;
}

export interface ProsperTransactionRecords {
  action: TransactionAction;
  created_at: Generated<Timestamp>;
  id: Int8;
  snapshot_id: Int8;
  updated_by: string;
}

export interface ProsperTransactionSnapshots {
  amount: Int8;
  created_at: Generated<Timestamp>;
  id: Int8;
  type: WalletType;
  updated_by_id: string;
  updated_by_name: string;
  wallet_category_id: number;
  wallet_currency_id: number;
  wallet_id: string;
  wallet_name: string;
  wallet_owners: Json;
}

export interface ProsperWalletAmounts {
  amount: Int8;
  created_at: Generated<Timestamp>;
  id: Int8;
  updated_at: Generated<Timestamp>;
  wallet_reference_id: Int8;
}

export interface ProsperWalletOwners {
  id: Int8;
  owner_type: OwnerType;
  profile_id: string;
  wallet_reference_id: Int8;
  wallet_type: WalletType;
}

export interface ProsperWalletReferences {
  budget_wallet_id: Int8 | null;
  id: Int8;
  savings_wallet_id: Int8 | null;
}

export interface DB {
  "auth.audit_log_entries": AuthAuditLogEntries;
  "auth.flow_state": AuthFlowState;
  "auth.identities": AuthIdentities;
  "auth.instances": AuthInstances;
  "auth.mfa_amr_claims": AuthMfaAmrClaims;
  "auth.mfa_challenges": AuthMfaChallenges;
  "auth.mfa_factors": AuthMfaFactors;
  "auth.one_time_tokens": AuthOneTimeTokens;
  "auth.refresh_tokens": AuthRefreshTokens;
  "auth.saml_providers": AuthSamlProviders;
  "auth.saml_relay_states": AuthSamlRelayStates;
  "auth.schema_migrations": AuthSchemaMigrations;
  "auth.sessions": AuthSessions;
  "auth.sso_domains": AuthSsoDomains;
  "auth.sso_providers": AuthSsoProviders;
  "auth.users": AuthUsers;
  "constants.categories": ConstantsCategories;
  "constants.currencies": ConstantsCurrencies;
  "prosper.budget_metadata": ProsperBudgetMetadata;
  "prosper.partner_invitations": ProsperPartnerInvitations;
  "prosper.profiles": ProsperProfiles;
  "prosper.savings_metadata": ProsperSavingsMetadata;
  "prosper.transaction_records": ProsperTransactionRecords;
  "prosper.transaction_snapshots": ProsperTransactionSnapshots;
  "prosper.wallet_amounts": ProsperWalletAmounts;
  "prosper.wallet_owners": ProsperWalletOwners;
  "prosper.wallet_references": ProsperWalletReferences;
}